+++ pkg/spinup/spinup.go
@@ -4,6 +4,8 @@ import (
 	"encoding/json"
 	"io/ioutil"
 	"net/http"
+	"net/url"
+	"strconv"
 
 	"github.com/pkg/errors"
 	log "github.com/sirupsen/logrus"
@@ -32,14 +34,27 @@ type ResourceType interface {
 	GetEndpoint(id string) string
 }
 
-// Resource gets details about a resource
+func New(spinupUrl string, client *http.Client) (*Client, error) {
+	u, err := url.Parse(spinupUrl)
+	if err != nil {
+		return nil, err
+	}
+
+	BaseURL = u.String()
+	return &Client{
+		// BaseURL:    u,
+		HTTPClient: client,
+	}, nil
+}
+
+// Resource gets details about a resource and unmarshals them them into the passed ResourceType
 func (c *Client) GetResource(id string, r ResourceType) error {
 	res, err := c.HTTPClient.Get(r.GetEndpoint(id))
 	if err != nil {
 		return errors.Wrap(err, "failed getting resource "+id)
 	}
 
-	if res.StatusCode > 400 {
+	if res.StatusCode >= 400 {
 		return errors.New("error getting resource details: " + res.Status)
 	}
 
@@ -61,4 +76,25 @@ func (c *Client) GetResource(id string, r ResourceType) error {
 	log.Debugf("decoded output: %+v", r)
 
 	return nil
+}
+
+func (fi *FlexInt) UnmarshalJSON(b []byte) error {
+	if b[0] != '"' {
+		return json.Unmarshal(b, (*int)(fi))
+	}
+	var s string
+	if err := json.Unmarshal(b, &s); err != nil {
+		return err
+	}
+	i, err := strconv.Atoi(s)
+	if err != nil {
+		return err
+	}
+	*fi = FlexInt(i)
+	return nil
+}
+
+func (fi *FlexInt) String() string {
+	log.Debugf("converting flex int to string: %v", *fi)
+	return strconv.Itoa(int(*fi))
 }
